Refactor into base class
